Moves vertical menu element to proper context namespace
