+++ src/Sylius/Behat/Resources/config/services/elements/product.xml
@@ -28,5 +28,7 @@
         <service id="sylius.behat.element.product.show.taxonomy" class="Sylius\Behat\Element\Product\ShowPage\TaxonomyElement" parent="sylius.behat.element" public="false" />
         <service id="sylius.behat.element.product.show.options" class="Sylius\Behat\Element\Product\ShowPage\OptionsElement" parent="sylius.behat.element" public="false" />
         <service id="sylius.behat.element.product.show.variants" class="Sylius\Behat\Element\Product\ShowPage\VariantsElement" parent="sylius.behat.element" public="false" />
+
+        <service id="sylius.behat.element.product.index.vertical_menu" class="Sylius\Behat\Element\Product\IndexPage\VerticalMenuElement" parent="sylius.behat.element" public="false" />
     </services>
 </container>
