+++ src/Sylius/Behat/Resources/config/services/elements/shop.xml
@@ -21,6 +21,5 @@
         <service id="sylius.behat.element.shop.account.register" class="Sylius\Behat\Element\Shop\Account\RegisterElement" parent="sylius.behat.element" public="false" />
 
         <service id="sylius.behat.element.shop.menu" class="Sylius\Behat\Element\Shop\MenuElement" parent="sylius.behat.element" public="false" />
-        <service id="sylius.behat.element.shop.vertical_menu" class="Sylius\Behat\Element\Shop\VerticalMenuElement" parent="sylius.behat.element" public="false" />
     </services>
 </container>
